Adding tweet vetting pre-DB insertion

+++ back-end/seed/test.seed.js
@@ -1,36 +1,50 @@
-const {Tweets, Users} = require('../models/models.js');
+const fs = require('fs');
+const { Tweets, Users } = require('../models/models.js');
 
 const testData = require('./exampleTweets.json');
 const testUsers = require('./exampleUsers.json');
+
 const mongoose = require('mongoose');
 mongoose.Promise = global.Promise;
-const config = require('../server/config')
+const config = require('../server/config');
+
+const translate = require('@google-cloud/translate')();
+
+const { syntaxOfTweet, normaliseTweet, selectWordType, translateWord, randomWords, pickCorrectWord } = require('../server/controllers/utils');
 
 
-function seedDataBase () {
-    
-    mongoose.connect(config.url, {useMongoClient: true})
 
-    Tweets.collection.drop()
-    Users.collection.drop()
+function seedDataBase(tweets, users) {
 
-    Promise.all(seedUsers())
-    .then(data => {
-        console.log('the collection of users was seeded with', data.length, 'users')
-        mongoose.disconnect()
-    })
-    .catch(console.error)
+    if (tweets === undefined && users === undefined) return;
 
-    Promise.all(seedTweets())
-    .then(data => {
-        console.log('the collection of tweets was seeded with', data.length, 'tweets')
-        mongoose.disconnect()
-    })
-    .catch(console.error)
+    mongoose.connect(config.url, { useMongoClient: true })
+
+    if (users) {
+        Users.collection.drop()
+
+        Promise.all(seedUsers())
+            .then(data => {
+                console.log('the collection of users was seeded with', data.length, 'users')
+                mongoose.disconnect()
+            })
+            .catch(console.error)
+    }
+
+    if (tweets) {
+        Tweets.collection.drop()
+
+        Promise.all(seedTweets())
+            .then(data => {
+                console.log('the collection of tweets was seeded with', data.length, 'tweets')
+                mongoose.disconnect()
+            })
+            .catch(console.error)
+    }
 }
 
-function seedUsers () {
-    
+function seedUsers() {
+
     return testUsers.map((user) => {
         return new Users({
             name: user.name,
@@ -38,22 +52,65 @@ function seedUsers () {
             avatar: user.avatar,
             completedTweets: user.completedTweets
         })
-        .save()
+            .save()
     });
 }
 
-function seedTweets () {
-    return testData.map((tweet) => {
-        return new Tweets({
-          created_at: tweet.created_at,
-          id: tweet.id,
-          text: tweet.text,
-          entities: tweet.entities,
-          user_screen_name: tweet.user.screen_name,
-          user_profile_image: tweet.user.profile_image_url 
+function seedTweets(vettedTweets) {
+
+    return vettedTweets.map((tweet) => {
+        return new Tweets(tweet).save()
+    });
+}
+
+
+function vetTweets() {
+    filteredTweets = [];
+
+    console.log("Unfiltered tweets: " + testData.length)
+
+    normalisedTweets = testData.map((tweet) => {
+        return {
+            created_at: tweet.created_at,
+            id: tweet.id,
+            text: tweet.text,
+            normalisedText: normaliseTweet(tweet),
+            entities: tweet.entities,
+            user_screen_name: tweet.user.screen_name,
+            user_profile_image: tweet.user.profile_image_url
+        }
+    })
+
+    let promiseArr = normalisedTweets.map((tweet) => {
+        return syntaxOfTweet(tweet.id + " " + tweet.normalisedText)
+    })
+
+    Promise.all(promiseArr)
+        .then((data) => {
+
+            filteredSyntaxArr = data.filter((tweetSyntax) => {
+                const syntaxArrayWithoutID = tweetSyntax.slice(1);
+                return selectWordType(syntaxArrayWithoutID, 'ADJ') !== null;
+            })
+
+            filteredSyntaxArr.forEach((syntaxArr) => {
+                normalisedTweets.forEach((tweet) => {
+                    if (syntaxArr[0].word == tweet.id) {
+                        filteredTweets.push(
+                            {
+                                tweet,
+                                wordArr: syntaxArr.slice(1)
+                            }
+                        )
+                    }
+                })
+            })
+
+            const removedTweets = testData.length - filteredTweets.length
+            console.log(`Removed ${removedTweets} tweet${removedTweets === 1 ? '' : 's'} leaving ${filteredTweets.length} vetted tweets`)
+
+            seedDataBase(seedTweets(filteredTweets), seedUsers())
         })
-        .save()
-    });
 }
-
-seedDataBase();
+vetTweets();
+// seedDataBase();
